feat(package): move functionality to donut-common
