+++ packages/common/src/index.js
@@ -1,4 +1,9 @@
 var rater = require('./rater')
+var keys = require('./keys.js')
+var getRandomDonuts = require('./get-random-donuts.js')
+
 module.exports = {
+  getRandomDonuts,
+  keys,
   rater
 }
