+++ packages/common/src/keys.js
@@ -1,0 +1,7 @@
+module.exports = [
+  'DONUT_FROSTING_COVERAGE',
+  'DONUT_FROSTING_THICKNESS',
+  'DONUT_SPRINKLE_COVERAGE',
+  'DONUT_INNER_RADIUS',
+  'DONUT_OUTER_RADIUS'
+]
