+++ packages/regression/package.json
@@ -32,8 +32,7 @@
   },
   "dependencies": {
     "bluebird": "^3.5.1",
-    "debug": "^3.1.0",
-    "lodash": "^4.17.4"
+    "debug": "^3.1.0"
   },
   "private": true
 }
